Toggle TEST_MODE

+++ uninstall.php
@@ -15,8 +15,21 @@
 		}
 	}
 
+	if (isset($_POST['toggle_test_mode'])) {
+		chmod(BASE_DIR . "/" . SETTINGS_INI, 0755);
+		$content = file_get_contents(BASE_DIR . "/" . SETTINGS_INI);
+		$value = 'true';
+		if (TEST_MODE) {
+			$value = 'false';
+		}
+		$content = preg_replace("#define\('TEST_MODE', (.*)\);#", "define('TEST_MODE', $value);", $content);
+		file_put_contents(BASE_DIR . "/" . SETTINGS_INI, $content);
+		chmod(BASE_DIR . "/" . SETTINGS_INI, 0400);
+		redirect_to('index.php');
+	}
+
 	if (isset($_POST['remove_settings_only'])) {
-			chmod(BASE_DIR . "/" . SETTINGS_INI, 0755);
+		chmod(BASE_DIR . "/" . SETTINGS_INI, 0755);
 		unlink(BASE_DIR . "/" . SETTINGS_INI);
 ?>
 <html>
@@ -62,6 +75,18 @@ EOF;
 	<form name="input" action="uninstall.php" method="POST">
 		<input type="hidden" name="remove_settings_only" value="yes"/>
 		<input type="submit" value="Submit"/>
+	</form>
+
+	Toggle TEST_MODE
+	<?php
+		$button = 'on';
+		if (TEST_MODE) {
+			$button = 'off';
+		}
+	?>
+	<form name="input" action="uninstall.php" method="POST">
+		<input type="hidden" name="toggle_test_mode" value="yes"/>
+		<input type="submit" value="Turn <?php echo $button; ?>"/>
 	</form>
 	<a href="support/save_db.php">Backup the database</a>
 </html>
